Crate objc2_io_kit

Bindings to the IOKit framework

See Apple’s docs and the general docs on framework crates for more information.

Structs

EvCmd

Apple’s documentation

IOAsyncCompletionContent

Apple’s documentation

IOCFPlugInInterfaceStruct

Apple’s documentation

IODataQueueEntry

Represents an entry within the data queue

IODataQueueMemory

A struct mapping to the header region of a data queue.

IOFixedPoint32

Apple’s documentation

IOHIDAccessType

Apple’s documentation

IOHIDButtonModes

Apple’s documentation

IOHIDEventSystemClient

IOHIDEventSystemClient

IOHIDRequestType

Apple’s documentation

IOHIDServiceClient

IOHIDServiceClient

IOHIDUserDevice

Apple’s documentation

IOHIDUserDeviceOptions

Enumerator of IOHIDUserDeviceOptions to be passed in to IOHIDUserDeviceCreateWithOptions.

IONamedValue

Apple’s documentation

IONotificationPort

Apple’s documentation

IOPhysicalRange

Apple’s documentation

IORPCMessage

Apple’s documentation

IORPCMessageErrorReturnContent

Apple’s documentation

IOServiceInterestContent

Apple’s documentation

IOURLError

Apple’s documentation

IOVirtualRange

Apple’s documentation

NXEvent

Apple’s documentation

NXEventData_compound

Apple’s documentation

NXEventData_key

Apple’s documentation

NXEventData_mouse

Apple’s documentation

NXEventData_mouseMove

Apple’s documentation

NXEventData_proximity

Apple’s documentation

NXEventData_scrollWheel

Apple’s documentation

NXEventData_tablet

Apple’s documentation

NXEventData_tablet_tilt

Apple’s documentation

NXEventData_tracking

Apple’s documentation

NXEventSystemDevice

Apple’s documentation

NXEventSystemDeviceList

Apple’s documentation

NXMouseButton

Apple’s documentation

NXPoint

Apple’s documentation

NXSize

Apple’s documentation

NXTabletPointData

Apple’s documentation

NXTabletProximityData

Apple’s documentation

OSClassDescription

Apple’s documentation

evsioEVSIOCCSIndices

Apple’s documentation

evsioEVSIOSCSIndices

Apple’s documentation

evsioKeymapping

Apple’s documentation

evsioMouseScaling

Apple’s documentation

Constants

EVENT_SYSTEM_VERSION

Apple’s documentation

IO_OBJECT_NULL

Apple’s documentation

kFirstIOKitNotificationType

Apple’s documentation

kHIDAccelGainCubicKey

Apple’s documentation

kHIDAccelGainLinearKey

Apple’s documentation

kHIDAccelGainParabolicKey

Apple’s documentation

kHIDAccelGainQuarticKey

Apple’s documentation

kHIDAccelIndexKey

Apple’s documentation

kHIDAccelParametricCurvesDebugKey

Apple’s documentation

kHIDAccelTangentSpeedCubicRootKey

Apple’s documentation

kHIDAccelTangentSpeedLinearKey

Apple’s documentation

kHIDAccelTangentSpeedParabolicRootKey

Apple’s documentation

kHIDAccelTangentSpeedQuarticRootKey

Apple’s documentation

kHIDScrollAccelParametricCurvesDebugKey

Apple’s documentation

kHIDTrackingAccelParametricCurvesKey

Apple’s documentation

kIOAllCPUInitializedKey

Apple’s documentation

kIOAppPowerStateInterest

Apple’s documentation

kIOAsyncCalloutCount

Apple’s documentation

kIOAsyncCalloutFuncIndex

Apple’s documentation

kIOAsyncCalloutRefconIndex

Apple’s documentation

kIOAsyncCompletionNotificationType

Apple’s documentation

kIOAsyncReservedCount

Apple’s documentation

kIOAsyncReservedIndex

Apple’s documentation

kIOAudioPlane

Apple’s documentation

kIOBSDKey

Apple’s documentation

kIOBSDMajorKey

Apple’s documentation

kIOBSDMinorKey

Apple’s documentation

kIOBSDNameKey

Apple’s documentation

kIOBSDNamesKey

Apple’s documentation

kIOBSDUnitKey

Apple’s documentation

kIOBootDeviceKey

Apple’s documentation

kIOBootDevicePathKey

Apple’s documentation

kIOBootDeviceSizeKey

Apple’s documentation

kIOBundleResourceFileKey

Apple’s documentation

kIOBusBadgeKey

Apple’s documentation

kIOBusyInterest

Apple’s documentation

kIOCFPlugInTypesKey

Apple’s documentation

kIOCFSerializeToBinary

Apple’s documentation

kIOCatalogAddDrivers

Apple’s documentation

kIOCatalogAddDriversNoMatch

Apple’s documentation

kIOCatalogGetCacheMissList

Apple’s documentation

kIOCatalogGetContents

Apple’s documentation

kIOCatalogGetModuleDemandList

Apple’s documentation

kIOCatalogGetROMMkextList

Apple’s documentation

kIOCatalogKextdActive

Apple’s documentation

kIOCatalogKextdFinishedLaunching

Apple’s documentation

kIOCatalogModuleTerminate

Apple’s documentation

kIOCatalogModuleUnload

Apple’s documentation

kIOCatalogRemoveDrivers

Apple’s documentation

kIOCatalogRemoveDriversNoMatch

Apple’s documentation

kIOCatalogRemoveKernelLinker__Removed

Apple’s documentation

kIOCatalogResetDefault

Apple’s documentation

kIOCatalogResetDrivers

Apple’s documentation

kIOCatalogResetDriversNoMatch

Apple’s documentation

kIOCatalogServiceTerminate

Apple’s documentation

kIOCatalogStartMatching__Removed

Apple’s documentation

kIOCatalogueKey

Apple’s documentation

kIOClassKey

Apple’s documentation

kIOCommandPoolSizeKey

Apple’s documentation

kIOCompatibilityMatchKey

Apple’s documentation

kIOCompatibilityPropertiesKey

Apple’s documentation

kIOConnectMethodVarOutputSize

Apple’s documentation

kIOCopybackCache

Apple’s documentation

kIOCopybackInnerCache

Apple’s documentation

kIODEXTMatchCountKey

Apple’s documentation

kIODTNVRAMPanicInfoKey

Apple’s documentation

kIODefaultCache

Apple’s documentation

kIODefaultMatchCategoryKey

Apple’s documentation

kIODefaultMemoryType

Apple’s documentation

kIODeviceIconKey

Apple’s documentation

kIODeviceTreePlane

Apple’s documentation

kIODriverKitEntitlementKey

Apple’s documentation

kIODriverKitHIDFamilyDeviceEntitlementKey

Apple’s documentation

kIODriverKitHIDFamilyEventServiceEntitlementKey

Apple’s documentation

kIODriverKitHIDTransportEntitlementKey

Apple’s documentation

kIODriverKitRequiredEntitlementsKey

Apple’s documentation

kIODriverKitTestDriverEntitlementKey

Apple’s documentation

kIODriverKitTransportBuiltinEntitlementKey

Apple’s documentation

kIODriverKitUSBTransportEntitlementKey

Apple’s documentation

kIODriverKitUserClientEntitlementAdministratorKey

Apple’s documentation

kIODriverKitUserClientEntitlementAllowAnyKey

Apple’s documentation

kIODriverKitUserClientEntitlementAllowThirdPartyUserClientsKey

Apple’s documentation

kIODriverKitUserClientEntitlementCommunicatesWithDriversKey

Apple’s documentation

kIODriverKitUserClientEntitlementsKey

Apple’s documentation

kIOFireWirePlane

Apple’s documentation

kIOFirstMatchNotification

Apple’s documentation

kIOFirstPublishNotification

Apple’s documentation

kIOGeneralInterest

Apple’s documentation

kIOHIDActivityDisplayOn

Apple’s documentation

kIOHIDActivityUserIdle

Apple’s documentation

kIOHIDActivityUserIdleKey

Apple’s documentation

kIOHIDAutoDimBrightnessKey

Apple’s documentation

kIOHIDAutoDimStateKey

Apple’s documentation

kIOHIDAutoDimThresholdKey

Apple’s documentation

kIOHIDAutoDimTimeKey

Apple’s documentation

kIOHIDBrightnessKey

Apple’s documentation

kIOHIDCapsLockState

Apple’s documentation

kIOHIDClickSpaceKey

Apple’s documentation

kIOHIDClickTimeKey

Apple’s documentation

kIOHIDCurrentShmemVersion

Apple’s documentation

kIOHIDDisallowRemappingOfPrimaryClickKey

Apple’s documentation

kIOHIDEventNotification

Apple’s documentation

kIOHIDEventQueueTypeKernel

Apple’s documentation

kIOHIDEventQueueTypeUser

Apple’s documentation

kIOHIDEventSystemConnectType

Apple’s documentation

kIOHIDF12EjectDelayKey

Apple’s documentation

kIOHIDFKeyModeKey

Apple’s documentation

kIOHIDGlobalMemory

Apple’s documentation

kIOHIDIdleTimeKey

Apple’s documentation

kIOHIDInitialKeyRepeatKey

Apple’s documentation

kIOHIDInterfaceIDKey

Apple’s documentation

kIOHIDKeyMappingKey

Apple’s documentation

kIOHIDKeyRepeatKey

Apple’s documentation

kIOHIDKeyboardCapsLockDoesLockKey

Apple’s documentation

kIOHIDKeyboardGlobalModifiersKey

Apple’s documentation

kIOHIDKeyboardModifierMappingDstKey

Apple’s documentation

kIOHIDKeyboardModifierMappingPairsKey

Apple’s documentation

kIOHIDKeyboardModifierMappingSrcKey

Apple’s documentation

kIOHIDKeyboardSupportedModifiersKey

Apple’s documentation

kIOHIDKeyboardSupportsF12EjectKey

Apple’s documentation

kIOHIDKindKey

Apple’s documentation

kIOHIDLastCompatibleShmemVersion

Apple’s documentation

kIOHIDMessageRelayServiceInterfaceActive

Apple’s documentation

kIOHIDMouseKeysEnablesVirtualNumPadKey

Apple’s documentation

kIOHIDMouseKeysOnKey

Apple’s documentation

kIOHIDMouseKeysOptionTogglesKey

Apple’s documentation

kIOHIDNumLockState

Apple’s documentation

kIOHIDOpenedByEventSystem

kIOHIDOpenedByEventSystem

kIOHIDOpenedByFastPathClient

kIOHIDOpenedByFastPathClient

kIOHIDOriginalSubinterfaceIDKey

Apple’s documentation

kIOHIDParamConnectType

Apple’s documentation

kIOHIDParametersKey

Apple’s documentation

kIOHIDPointerAccelerationSettingsKey

Apple’s documentation

kIOHIDPointerAccelerationTableKey

Apple’s documentation

kIOHIDPointerButtonCountKey

Apple’s documentation

kIOHIDPointerContactToMoveKey

Apple’s documentation

kIOHIDPointerConvertAbsoluteKey

Apple’s documentation

kIOHIDPointerPressureToClickKey

Apple’s documentation

kIOHIDPointerResolutionKey

Apple’s documentation

kIOHIDPostHIDManagerEvent

Apple’s documentation

kIOHIDResetKeyboardKey

Apple’s documentation

kIOHIDResetLEDsKey

Apple’s documentation

kIOHIDResetPointerKey

Apple’s documentation

kIOHIDResetStickyKeyNotification

Apple’s documentation

kIOHIDScrollAccelerationTableKey

Apple’s documentation

kIOHIDScrollAccelerationTableXKey

Apple’s documentation

kIOHIDScrollAccelerationTableYKey

Apple’s documentation

kIOHIDScrollAccelerationTableZKey

Apple’s documentation

kIOHIDScrollCountAccelerationFactorKey

Apple’s documentation

kIOHIDScrollCountBootDefaultKey

Apple’s documentation

kIOHIDScrollCountIgnoreMomentumScrollsKey

Apple’s documentation

kIOHIDScrollCountMaxKey

Apple’s documentation

kIOHIDScrollCountMaxTimeDeltaBetweenKey

Apple’s documentation

kIOHIDScrollCountMaxTimeDeltaToSustainKey

Apple’s documentation

kIOHIDScrollCountMinDeltaToStartKey

Apple’s documentation

kIOHIDScrollCountMinDeltaToSustainKey

Apple’s documentation

kIOHIDScrollCountMouseCanResetKey

Apple’s documentation

kIOHIDScrollCountResetKey

Apple’s documentation

kIOHIDScrollCountZeroKey

Apple’s documentation

kIOHIDScrollMouseButtonKey

Apple’s documentation

kIOHIDScrollResetKey

Apple’s documentation

kIOHIDScrollZoomModifierMaskKey

Apple’s documentation

kIOHIDServerConnectType

Apple’s documentation

kIOHIDServiceGlobalModifiersUsageKey

Apple’s documentation

kIOHIDSetCursorPosition

Apple’s documentation

kIOHIDSetGlobalEventFlags

Apple’s documentation

kIOHIDSetRelativeCursorPosition

Apple’s documentation

kIOHIDSlowKeysDelayKey

Apple’s documentation

kIOHIDStickyKeysDisabledKey

Apple’s documentation

kIOHIDStickyKeysOnKey

Apple’s documentation

kIOHIDStickyKeysShiftTogglesKey

Apple’s documentation

kIOHIDSubinterfaceIDKey

Apple’s documentation

kIOHIDSystemClass

Apple’s documentation

kIOHIDUseKeyswitchKey

Apple’s documentation

kIOHIDVirtualHIDevice

Apple’s documentation

kIOHIDWaitCursorFrameIntervalKey

Apple’s documentation

kIOHIKeyboardClass

Apple’s documentation

kIOHIPointingClass

Apple’s documentation

kIOIconKey

Apple’s documentation

kIOInhibitCache

Apple’s documentation

kIOInterestCalloutCount

Apple’s documentation

kIOInterestCalloutFuncIndex

Apple’s documentation

kIOInterestCalloutRefconIndex

Apple’s documentation

kIOInterestCalloutServiceIndex

Apple’s documentation

kIOKitBuildVersionKey

Apple’s documentation

kIOKitDebugKey

Apple’s documentation

kIOKitDiagnosticsKey

Apple’s documentation

kIOKitNoticationMsgSizeMask

Apple’s documentation

kIOKitNoticationTypeMask

Apple’s documentation

kIOKitNoticationTypeSizeAdjShift

Apple’s documentation

kIOLocationMatchKey

Apple’s documentation

kIOMapAnywhere

Apple’s documentation

kIOMapCacheMask

Apple’s documentation

kIOMapCacheShift

Apple’s documentation

kIOMapCopybackCache

Apple’s documentation

kIOMapCopybackInnerCache

Apple’s documentation

kIOMapDefaultCache

Apple’s documentation

kIOMapGuardedLarge

Apple’s documentation

kIOMapGuardedMask

Apple’s documentation

kIOMapGuardedSmall

Apple’s documentation

kIOMapInhibitCache

Apple’s documentation

kIOMapOverwrite

Apple’s documentation

kIOMapPostedCombinedReordered

Apple’s documentation

kIOMapPostedReordered

Apple’s documentation

kIOMapPostedWrite

Apple’s documentation

kIOMapPrefault

Apple’s documentation

kIOMapReadOnly

Apple’s documentation

kIOMapRealTimeCache

Apple’s documentation

kIOMapReference

Apple’s documentation

kIOMapStatic

Apple’s documentation

kIOMapUnique

Apple’s documentation

kIOMapUserOptionsMask

Apple’s documentation

kIOMapWriteCombineCache

Apple’s documentation

kIOMapWriteThruCache

Apple’s documentation

kIOMapperIDKey

Apple’s documentation

kIOMatchCategoryKey

Apple’s documentation

kIOMatchDeferKey

Apple’s documentation

kIOMatchedNotification

Apple’s documentation

kIOMatchedPersonalityKey

Apple’s documentation

kIOMatchedServiceCountKey

Apple’s documentation

kIOMatchingCalloutCount

Apple’s documentation

kIOMatchingCalloutFuncIndex

Apple’s documentation

kIOMatchingCalloutRefconIndex

Apple’s documentation

kIOMaxBusStall5usec

Apple’s documentation

kIOMaxBusStall10usec

Apple’s documentation

kIOMaxBusStall20usec

Apple’s documentation

kIOMaxBusStall25usec

Apple’s documentation

kIOMaxBusStall30usec

Apple’s documentation

kIOMaxBusStall40usec

Apple’s documentation

kIOMaxBusStallNone

Apple’s documentation

kIOMaximumBlockCountReadKey

Apple’s documentation

kIOMaximumBlockCountWriteKey

Apple’s documentation

kIOMaximumByteCountReadKey

Apple’s documentation

kIOMaximumByteCountWriteKey

Apple’s documentation

kIOMaximumPriorityCountKey

Apple’s documentation

kIOMaximumSegmentAddressableBitCountKey

Apple’s documentation

kIOMaximumSegmentByteCountReadKey

Apple’s documentation

kIOMaximumSegmentByteCountWriteKey

Apple’s documentation

kIOMaximumSegmentCountReadKey

Apple’s documentation

kIOMaximumSegmentCountWriteKey

Apple’s documentation

kIOMaximumSwapWriteKey

Apple’s documentation

kIOMessageCanDevicePowerOff

Apple’s documentation

kIOMessageCanSystemPowerOff

Apple’s documentation

kIOMessageCanSystemSleep

Apple’s documentation

kIOMessageConsoleSecurityChange

Apple’s documentation

kIOMessageCopyClientID

Apple’s documentation

kIOMessageDeviceHasPoweredOff

Apple’s documentation

kIOMessageDeviceHasPoweredOn

Apple’s documentation

kIOMessageDeviceSignaledWakeup

Apple’s documentation

kIOMessageDeviceWillNotPowerOff

Apple’s documentation

kIOMessageDeviceWillPowerOff

Apple’s documentation

kIOMessageDeviceWillPowerOn

Apple’s documentation

kIOMessageServiceBusyStateChange

Apple’s documentation

kIOMessageServiceIsAttemptingOpen

Apple’s documentation

kIOMessageServiceIsRequestingClose

Apple’s documentation

kIOMessageServiceIsResumed

Apple’s documentation

kIOMessageServiceIsSuspended

Apple’s documentation

kIOMessageServiceIsTerminated

Apple’s documentation

kIOMessageServicePropertyChange

Apple’s documentation

kIOMessageServiceWasClosed

Apple’s documentation

kIOMessageSystemCapabilityChange

Apple’s documentation

kIOMessageSystemHasPoweredOn

Apple’s documentation

kIOMessageSystemPagingOff

Apple’s documentation

kIOMessageSystemWillNotPowerOff

Apple’s documentation

kIOMessageSystemWillNotSleep

Apple’s documentation

kIOMessageSystemWillPowerOff

Apple’s documentation

kIOMessageSystemWillPowerOn

Apple’s documentation

kIOMessageSystemWillRestart

Apple’s documentation

kIOMessageSystemWillSleep

Apple’s documentation

kIOMinimumSaturationByteCountKey

Apple’s documentation

kIOMinimumSegmentAlignmentByteCountKey

Apple’s documentation

kIONVRAMActivateCSRConfigPropertyKey

Apple’s documentation

kIONVRAMBootArgsKey

Apple’s documentation

kIONVRAMDeletePropertyKey

Apple’s documentation

kIONVRAMDeletePropertyKeyWRet

Apple’s documentation

kIONVRAMReadAccessKey

Apple’s documentation

kIONVRAMSyncNowPropertyKey

Apple’s documentation

kIONVRAMSystemAllowKey

Apple’s documentation

kIONVRAMWriteAccessKey

Apple’s documentation

kIONameMatchKey

Apple’s documentation

kIONameMatchedKey

Apple’s documentation

kIOPSFamilyCodeUnsupported

Apple’s documentation

kIOParentMatchKey

Apple’s documentation

kIOPathKey

Apple’s documentation

kIOPathMatchKey

Apple’s documentation

kIOPlatformDeviceMessageKey

Apple’s documentation

kIOPlatformSerialNumberKey

Apple’s documentation

kIOPlatformUUIDKey

Apple’s documentation

kIOPostedCombinedReordered

Apple’s documentation

kIOPostedReordered

Apple’s documentation

kIOPostedWrite

Apple’s documentation

kIOPowerPlane

Apple’s documentation

kIOPriorityPowerStateInterest

Apple’s documentation

kIOProbeScoreKey

Apple’s documentation

kIOPropertyExistsMatchKey

Apple’s documentation

kIOPropertyMatchKey

Apple’s documentation

kIOProviderClassKey

Apple’s documentation

kIOPublishNotification

Apple’s documentation

kIORPCMessageError

Apple’s documentation

kIORPCMessageKernel

Apple’s documentation

kIORPCMessageLocalHost

Apple’s documentation

kIORPCMessageObjectRefs

Apple’s documentation

kIORPCMessageOneway

Apple’s documentation

kIORPCMessageOnqueue

Apple’s documentation

kIORPCMessageRemote

Apple’s documentation

kIORPCMessageSimpleReply

Apple’s documentation

kIORPCVersion190615

Apple’s documentation

kIORPCVersion190615Reply

Apple’s documentation

kIORealTimeCache

Apple’s documentation

kIORegistryEntryAllowableSetPropertiesKey

Apple’s documentation

kIORegistryEntryDefaultLockingSetPropertiesKey

Apple’s documentation

kIORegistryEntryIDKey

Apple’s documentation

kIORegistryEntryPropertyKeysKey

Apple’s documentation

kIORegistryIterateParents

Apple’s documentation

kIORegistryIterateRecursively

Apple’s documentation

kIORegistryPlanesKey

Apple’s documentation

kIORematchCountKey

Apple’s documentation

kIORematchPersonalityKey

Apple’s documentation

kIOResourceMatchKey

Apple’s documentation

kIOResourceMatchedKey

Apple’s documentation

kIOResourcesClass

Apple’s documentation

kIOResourcesSetPropertyKey

Apple’s documentation

kIOReturnAborted

Apple’s documentation

kIOReturnBadArgument

Apple’s documentation

kIOReturnBadMedia

Apple’s documentation

kIOReturnBadMessageID

Apple’s documentation

kIOReturnBusy

Apple’s documentation

kIOReturnCannotLock

Apple’s documentation

kIOReturnCannotWire

Apple’s documentation

kIOReturnDMAError

Apple’s documentation

kIOReturnDeviceError

Apple’s documentation

kIOReturnError

Apple’s documentation

kIOReturnExclusiveAccess

Apple’s documentation

kIOReturnIOError

Apple’s documentation

kIOReturnIPCError

Apple’s documentation

kIOReturnInternalError

Apple’s documentation

kIOReturnInvalid

Apple’s documentation

kIOReturnIsoTooNew

Apple’s documentation

kIOReturnIsoTooOld

Apple’s documentation

kIOReturnLockedRead

Apple’s documentation

kIOReturnLockedWrite

Apple’s documentation

kIOReturnMessageTooLarge

Apple’s documentation

kIOReturnNoBandwidth

Apple’s documentation

kIOReturnNoChannels

Apple’s documentation

kIOReturnNoCompletion

Apple’s documentation

kIOReturnNoDevice

Apple’s documentation

kIOReturnNoFrames

Apple’s documentation

kIOReturnNoInterrupt

Apple’s documentation

kIOReturnNoMedia

Apple’s documentation

kIOReturnNoMemory

Apple’s documentation

kIOReturnNoPower

Apple’s documentation

kIOReturnNoResources

Apple’s documentation

kIOReturnNoSpace

Apple’s documentation

kIOReturnNotAligned

Apple’s documentation

kIOReturnNotAttached

Apple’s documentation

kIOReturnNotFound

Apple’s documentation

kIOReturnNotOpen

Apple’s documentation

kIOReturnNotPermitted

Apple’s documentation

kIOReturnNotPrivileged

Apple’s documentation

kIOReturnNotReadable

Apple’s documentation

kIOReturnNotReady

Apple’s documentation

kIOReturnNotResponding

Apple’s documentation

kIOReturnNotWritable

Apple’s documentation

kIOReturnOffline

Apple’s documentation

kIOReturnOverrun

Apple’s documentation

kIOReturnPortExists

Apple’s documentation

kIOReturnRLDError

Apple’s documentation

kIOReturnStillOpen

Apple’s documentation

kIOReturnSuccess

Apple’s documentation

kIOReturnTimeout

Apple’s documentation

kIOReturnUnderrun

Apple’s documentation

kIOReturnUnformattedMedia

Apple’s documentation

kIOReturnUnsupported

Apple’s documentation

kIOReturnUnsupportedMode

Apple’s documentation

kIOReturnVMError

Apple’s documentation

kIOServiceClass

Apple’s documentation

kIOServiceDEXTEntitlementsKey

Apple’s documentation

kIOServiceInteractionAllowed

Apple’s documentation

kIOServiceMatchedNotificationType

Apple’s documentation

kIOServiceMessageNotificationType

Apple’s documentation

kIOServicePlane

Apple’s documentation

kIOServicePublishNotificationType

Apple’s documentation

kIOServiceTerminatedNotificationType

Apple’s documentation

kIOStateNotificationEntitlementGetKey

Apple’s documentation

kIOStateNotificationEntitlementSetKey

Apple’s documentation

kIOStateNotificationItemCopyKey

Apple’s documentation

kIOStateNotificationItemCreateKey

Apple’s documentation

kIOStateNotificationItemSetKey

Apple’s documentation

kIOStateNotificationNameKey

Apple’s documentation

kIOSupportedPropertiesKey

Apple’s documentation

kIOSystemStateClamshellKey

Apple’s documentation

kIOSystemStateHaltDescriptionHaltStateKey

Apple’s documentation

kIOSystemStateHaltDescriptionKey

Apple’s documentation

kIOSystemStatePowerSourceDescriptionACAttachedKey

Apple’s documentation

kIOSystemStatePowerSourceDescriptionKey

Apple’s documentation

kIOSystemStateSleepDescriptionHibernateStateHibernating

Apple’s documentation

kIOSystemStateSleepDescriptionHibernateStateInactive

Apple’s documentation

kIOSystemStateSleepDescriptionHibernateStateKey

Apple’s documentation

kIOSystemStateSleepDescriptionHibernateStateWakingFromHibernate

Apple’s documentation

kIOSystemStateSleepDescriptionKey

Apple’s documentation

kIOSystemStateSleepDescriptionReasonKey

Apple’s documentation

kIOSystemStateWakeDescriptionContinuousTimeOffsetKey

Apple’s documentation

kIOSystemStateWakeDescriptionKey

Apple’s documentation

kIOSystemStateWakeDescriptionWakeReasonKey

Apple’s documentation

kIOTerminatedNotification

Apple’s documentation

kIOUSBPlane

Apple’s documentation

kIOUserClassKey

Apple’s documentation

kIOUserClassesKey

Apple’s documentation

kIOUserClientClassKey

Apple’s documentation

kIOUserClientCreatorKey

Apple’s documentation

kIOUserClientCrossEndianCompatibleKey

Apple’s documentation

kIOUserClientCrossEndianKey

Apple’s documentation

kIOUserClientDefaultLockingKey

Apple’s documentation

kIOUserClientDefaultLockingSetPropertiesKey

Apple’s documentation

kIOUserClientDefaultLockingSingleThreadExternalMethodKey

Apple’s documentation

kIOUserClientEntitlementsKey

Apple’s documentation

kIOUserClientSharedInstanceKey

Apple’s documentation

kIOUserServerClassKey

Apple’s documentation

kIOUserServerNameKey

Apple’s documentation

kIOUserServerOneProcessKey

Apple’s documentation

kIOUserServerPreserveUserspaceRebootKey

Apple’s documentation

kIOUserServerTagKey

Apple’s documentation

kIOUserServicePropertiesKey

Apple’s documentation

kIOUserUserClientKey

Apple’s documentation

kIOWillTerminateNotification

Apple’s documentation

kIOWriteCombineCache

Apple’s documentation

kIOWriteThruCache

Apple’s documentation

kLastIOKitNotificationType

Apple’s documentation

kMaxAsyncArgs

Apple’s documentation

kMicrosecondScale

Apple’s documentation

kMillisecondScale

Apple’s documentation

kNXEventDataVersion

Apple’s documentation

kNXEventVersion

Apple’s documentation

kNanosecondScale

Apple’s documentation

kOSAsyncCompleteMessageID

Apple’s documentation

kOSAsyncRefCount

Apple’s documentation

kOSAsyncRefSize

Apple’s documentation

kOSBuildVersionKey

Apple’s documentation

kOSClassCanRemote

Apple’s documentation

kOSNotificationMessageID

Apple’s documentation

kSecondScale

Apple’s documentation

kTickScale

Apple’s documentation

sub_iokit_acpi

Apple’s documentation

sub_iokit_acpiec

Apple’s documentation

sub_iokit_ahci

Apple’s documentation

sub_iokit_apfs

Apple’s documentation

sub_iokit_block_storage

Apple’s documentation

sub_iokit_bluetooth

Apple’s documentation

sub_iokit_common

Apple’s documentation

sub_iokit_firewire

Apple’s documentation

sub_iokit_graphics

Apple’s documentation

sub_iokit_graphics_acceleration

Apple’s documentation

sub_iokit_hidsystem

Apple’s documentation

sub_iokit_keystore

Apple’s documentation

sub_iokit_networking

Apple’s documentation

sub_iokit_pmu

Apple’s documentation

sub_iokit_powermanagement

Apple’s documentation

sub_iokit_scsi

Apple’s documentation

sub_iokit_smbus

Apple’s documentation

sub_iokit_timesync_avb

Apple’s documentation

sub_iokit_usb

Apple’s documentation

sub_iokit_usbaudio

Apple’s documentation

sub_iokit_wirelesscharging

Apple’s documentation

Statics

kIOMainPortDefault⚠

The default mach port used to initiate communication with IOKit.

kIOMasterPortDefault⚠Deprecated

Deprecated name for kIOMainPortDefault.

Functions

IOBSDNameMatching⚠

Create a matching dictionary that specifies an IOService match based on BSD device name.

IOCFSerialize⚠

Safety

IOCFUnserialize⚠

Safety

IOCFUnserializeBinary⚠

Safety

IOCFUnserializeWithSize⚠

Safety

IOCatalogueGetData⚠

Safety

IOCatalogueModuleLoaded⚠

Safety

IOCatalogueReset

IOCatalogueSendData⚠

Safety

IOCatalogueTerminate⚠

Safety

IOConnectAddClient

Inform a connection of a second connection.

IOConnectAddRef

Adds a reference to the connect handle.

IOConnectCallAsyncMethod⚠

Safety

IOConnectCallAsyncScalarMethod⚠

Safety

IOConnectCallAsyncStructMethod⚠

Safety

IOConnectCallMethod⚠

Safety

IOConnectCallScalarMethod⚠

Safety

IOConnectCallStructMethod⚠

Safety

IOConnectGetService⚠

Returns the IOService a connect handle was opened on.

IOConnectMapMemory⚠

Safety

IOConnectMapMemory64⚠

Map hardware or shared memory into the caller’s task.

IOConnectRelease

Remove a reference to the connect handle.

IOConnectSetCFProperties⚠

Set CF container based properties on a connection.

IOConnectSetCFProperty⚠

Set a CF container based property on a connection.

IOConnectSetNotificationPort

Set a port to receive family specific notifications.

IOConnectTrap0

IOConnectTrap1

IOConnectTrap2

IOConnectTrap3

IOConnectTrap4

IOConnectTrap5

IOConnectTrap6

IOConnectUnmapMemory⚠

IOConnectUnmapMemory64⚠

Remove a mapping made with IOConnectMapMemory64.

IOCreatePlugInInterfaceForService⚠

Safety

IOCreateReceivePort⚠

Creates and returns a mach port suitable for receiving IOKit messages of the specified type.

IODataQueueAllocateNotificationPort

Allocates and returns a new mach port able to receive data available notifications from an IODataQueue.

IODataQueueDataAvailable⚠

Used to determine if more data is available on the queue.

IODataQueueDequeue⚠

Dequeues the next available entry on the queue and copies it into the given data pointer.

IODataQueueEnqueue⚠

Enqueues a new entry on the queue.

IODataQueuePeek⚠

Used to peek at the next entry on the queue.

IODataQueueSetNotificationPort⚠

Creates a simple mach message targeting the mach port specified in port.

IODataQueueWaitForAvailableData⚠

Wait for an incoming dataAvailable message on the given notifyPort.

IODestroyPlugInInterface⚠

Safety

IOHIDCheckAccess

Checks if the process has access to a specific IOHIDRequestType. A process may request access by calling the IOHIDRequestAccess function.

IOHIDCopyCFTypeParameter⚠

Safety

IOHIDCreateSharedMemory

IOHIDEventSystemClientCopyPropertyDeprecated

IOHIDEventSystemClientCopyServicesDeprecated

IOHIDEventSystemClientCreateSimpleClientDeprecated

IOHIDEventSystemClientSetProperty⚠Deprecated

IOHIDGetAccelerationWithKey⚠Deprecated

Safety

IOHIDGetActivityState⚠Deprecated

Safety

IOHIDGetButtonEventNum⚠Deprecated

Safety

IOHIDGetModifierLockState⚠

Safety

IOHIDGetMouseAcceleration⚠Deprecated

Safety

IOHIDGetMouseButtonMode⚠Deprecated

Safety

IOHIDGetParameter⚠Deprecated

Safety

IOHIDGetScrollAcceleration⚠Deprecated

Safety

IOHIDGetStateForSelector⚠

Safety

IOHIDRegisterVirtualDisplay⚠Deprecated

Safety

IOHIDRequestAccess

Requests access from the user for a specific IOHIDRequestType.

IOHIDServiceClientConformsToDeprecated

IOHIDServiceClientCopyPropertyDeprecated

IOHIDServiceClientGetRegistryIDDeprecated

IOHIDServiceClientSetProperty⚠Deprecated

IOHIDSetAccelerationWithKey⚠Deprecated

Safety

IOHIDSetCFTypeParameter⚠

Safety

IOHIDSetCursorEnableDeprecated

IOHIDSetEventsEnable

IOHIDSetModifierLockState

IOHIDSetMouseAccelerationDeprecated

IOHIDSetMouseButtonModeDeprecated

IOHIDSetMouseLocationDeprecated

IOHIDSetParameter⚠Deprecated

Safety

IOHIDSetScrollAccelerationDeprecated

IOHIDSetStateForSelector

IOHIDUnregisterVirtualDisplayDeprecated

IOHIDUserDeviceActivateDeprecated

IOHIDUserDeviceCancelDeprecated

IOHIDUserDeviceCopyPropertyDeprecated

IOHIDUserDeviceCreateWithProperties⚠Deprecated

IOHIDUserDeviceHandleReportWithTimeStamp⚠Deprecated

IOHIDUserDeviceSetProperty⚠Deprecated

IOIteratorIsValid

Checks an iterator is still valid.

IOIteratorNext

Returns the next object in an iteration.

IOIteratorReset

Resets an iteration back to the beginning.

IOKitGetBusyState⚠

Returns the busyState of all IOServices.

IOMainPort⚠

Returns the mach port used to initiate communication with IOKit.

IOMasterPort⚠Deprecated

Deprecated name for IOMainPort().

IONotificationPortCreateDeprecated

IONotificationPortDestroy⚠Deprecated

IONotificationPortGetMachPort⚠Deprecated

IONotificationPortGetRunLoopSource⚠Deprecated

IONotificationPortSetImportanceReceiver⚠Deprecated

IOObjectConformsTo⚠

Performs an OSDynamicCast operation on an IOKit object.

IOObjectCopyBundleIdentifierForClass⚠

Return the bundle identifier of the given class.

IOObjectCopyClass

Return the class name of an IOKit object.

IOObjectCopySuperclassForClass⚠

Return the superclass name of the given class.

IOObjectGetClass⚠

Return the class name of an IOKit object.

IOObjectGetKernelRetainCount

Returns kernel retain count of an IOKit object.

IOObjectGetRetainCount

Returns kernel retain count of an IOKit object. Identical to IOObjectGetKernelRetainCount() but available prior to Mac OS 10.6.

IOObjectGetUserRetainCount

Returns the retain count for the current process of an IOKit object.

IOObjectIsEqualTo

Checks two object handles to see if they represent the same kernel object.

IOObjectRelease

Releases an object handle previously returned by IOKitLib.

IOObjectRetain

Retains an object handle previously returned by IOKitLib.

IOOpenFirmwarePathMatching⚠Deprecated

Safety

IORegistryCreateIterator⚠

Create an iterator rooted at the registry root.

IORegistryEntryCopyFromPath⚠

Looks up a registry entry by path.

IORegistryEntryCopyPath⚠

Create a path for a registry entry.

IORegistryEntryCreateCFProperties⚠

Create a CF dictionary representation of a registry entry’s property table.

IORegistryEntryCreateCFProperty⚠

Create a CF representation of a registry entry’s property.

IORegistryEntryCreateIterator⚠

Create an iterator rooted at a given registry entry.

IORegistryEntryFromPath⚠

Looks up a registry entry by path.

IORegistryEntryGetChildEntry⚠

Returns the first child of a registry entry in a plane.

IORegistryEntryGetChildIterator⚠

Returns an iterator over an registry entry’s child entries in a plane.

IORegistryEntryGetLocationInPlane⚠

Returns a C-string location assigned to a registry entry, in a specified plane.

IORegistryEntryGetName⚠

Returns a C-string name assigned to a registry entry.

IORegistryEntryGetNameInPlane⚠

Returns a C-string name assigned to a registry entry, in a specified plane.

IORegistryEntryGetParentEntry⚠

Returns the first parent of a registry entry in a plane.

IORegistryEntryGetParentIterator⚠

Returns an iterator over an registry entry’s parent entries in a plane.

IORegistryEntryGetPath⚠

Create a path for a registry entry.

IORegistryEntryGetProperty⚠

Safety

IORegistryEntryGetRegistryEntryID⚠

Returns an ID for the registry entry that is global to all tasks.

IORegistryEntryIDMatching⚠

Create a matching dictionary that specifies an IOService match based on a registry entry ID.

IORegistryEntryInPlane⚠

Determines if the registry entry is attached in a plane.

IORegistryEntrySearchCFProperty⚠

Create a CF representation of a registry entry’s property.

IORegistryEntrySetCFProperties⚠

Set CF container based properties in a registry entry.

IORegistryEntrySetCFProperty⚠

Set a CF container based property in a registry entry.

IORegistryGetRootEntry

Return a handle to the registry root.

IORegistryIteratorEnterEntry

Recurse into the current entry in the registry iteration.

IORegistryIteratorExitEntry

Exits a level of recursion, restoring the current entry.

IOServiceAddInterestNotification⚠

Register for notification of state changes in an IOService.

IOServiceAddMatchingNotification⚠

Look up registered IOService objects that match a matching dictionary, and install a notification request of new IOServices that match.

IOServiceAddNotification⚠Deprecated

Safety

IOServiceAuthorize

Authorize access to an IOService.

IOServiceClose

Close a connection to an IOService and destroy the connect handle.

IOServiceGetBusyState⚠

Returns the busyState of an IOService.

IOServiceGetMatchingService⚠

Look up a registered IOService object that matches a matching dictionary.

IOServiceGetMatchingServices⚠

Look up registered IOService objects that match a matching dictionary.

IOServiceMatchPropertyTable⚠

Match an IOService objects with matching dictionary.

IOServiceMatching⚠

Create a matching dictionary that specifies an IOService class match.

IOServiceNameMatching⚠

Create a matching dictionary that specifies an IOService name match.

IOServiceOFPathToBSDName⚠Deprecated

Safety

IOServiceOpen⚠

A request to create a connection to an IOService.

IOServiceOpenAsFileDescriptor

IOServiceRequestProbe

A request to rescan a bus for device changes.

IOURLCreateDataAndPropertiesFromResource⚠

Safety

IOURLCreatePropertyFromResource⚠

Safety

IOURLWriteDataAndPropertiesToResource⚠

Safety

NXClickTimeDeprecated

NXCloseEventStatusDeprecated

NXEventSystemInfo⚠Deprecated

Safety

NXGetClickSpace⚠Deprecated

Safety

NXKeyRepeatIntervalDeprecated

NXKeyRepeatThresholdDeprecated

NXOpenEventStatusDeprecated

NXResetKeyboardDeprecated

NXResetMouseDeprecated

NXSetClickSpace⚠Deprecated

Safety

NXSetClickTimeDeprecated

NXSetKeyRepeatIntervalDeprecated

NXSetKeyRepeatThresholdDeprecated

Type Aliases

EvOffsets

IOAddressRange

Apple’s documentation

IOAlignment

Apple’s documentation

IOAsyncCallback

standard callback function for asynchronous I/O requests with lots of extra arguments beyond a refcon and result code.

IOAsyncCallback0

standard callback function for asynchronous I/O requests with no extra arguments beyond a refcon and result code.

IOAsyncCallback1

standard callback function for asynchronous I/O requests with one extra argument beyond a refcon and result code. This is often a count of the number of bytes transferred

IOAsyncCallback2

standard callback function for asynchronous I/O requests with two extra arguments beyond a refcon and result code.

IOByteCount

Apple’s documentation

IOByteCount32

Apple’s documentation

IOByteCount64

Apple’s documentation

IOCFPlugInInterface

Apple’s documentation

IOCacheMode

Apple’s documentation

IODeviceNumber

Apple’s documentation

IOFixed

Apple’s documentation

IOItemCount

Apple’s documentation

IOLogicalAddress

Apple’s documentation

IOMessage

IOMessage.h

IONotificationPortRef

IOKitLib IOKitLib implements non-kernel task access to common IOKit object types - IORegistryEntry, IOService, IOIterator etc. These functions are generic - families may provide API that is more specific.
IOKitLib represents IOKit objects outside the kernel with the types io_object_t, io_registry_entry_t, io_service_t, & io_connect_t. Function names usually begin with the type of object they are compatible with - eg. IOObjectRelease can be used with any io_object_t. Inside the kernel, the c++ class hierarchy allows the subclasses of each object type to receive the same requests from user level clients, for example in the kernel, IOService is a subclass of IORegistryEntry, which means any of the IORegistryEntryXXX functions in IOKitLib may be used with io_service_t’s as well as io_registry_t’s. There are functions available to introspect the class of the kernel object which any io_object_t et al. represents. IOKit objects returned by all functions should be released with IOObjectRelease.

IOOptionBits

Apple’s documentation

IOPhysicalAddress

Apple’s documentation

IOPhysicalAddress32

Apple’s documentation

IOPhysicalAddress64

Apple’s documentation

IOPhysicalLength

Apple’s documentation

IOPhysicalLength32

Apple’s documentation

IOPhysicalLength64

Apple’s documentation

IOReturn

Apple’s documentation

IOServiceInterestCallback

Callback function to be notified of changes in state of an IOService.

IOServiceMatchingCallback

Callback function to be notified of IOService publication.

IOVersion

Apple’s documentation

IOVirtualAddress

Apple’s documentation

NXCoord

Apple’s documentation

NXEventHandle

Apple’s documentation

NXEventSystemInfoType

Apple’s documentation

NXKeyMapping

Apple’s documentation

NXMouseScaling

Apple’s documentation

NXParsedKeyMapping

Apple’s documentation

OSObjectRef

Apple’s documentation

io_connect_t

Apple’s documentation

io_enumerator_t

Apple’s documentation

io_ident_t

Apple’s documentation

io_iterator_t

Apple’s documentation

io_object_t

Apple’s documentation

io_registry_entry_t

Apple’s documentation

io_service_t

Apple’s documentation

uext_object_t

Apple’s documentation

Unions

NXEventData

Apple’s documentation

NXEventData_compound_misc

Apple’s documentation

NXEventData_mouseMove_tablet

Apple’s documentation

NXEventData_mouse_tablet

Apple’s documentation