objc2_io_kit

Function IORegistryCreateIterator

pub unsafe extern "C-unwind" fn IORegistryCreateIterator(
    main_port: mach_port_t,
    plane: *mut [c_char; 128],
    options: IOOptionBits,
    iterator: *mut io_iterator_t,
) -> kern_return_t

Create an iterator rooted at the registry root.

This method creates an IORegistryIterator in the kernel that is set up with options to iterate children of the registry root entry, and to recurse automatically into entries as they are returned, or only when instructed with calls to IORegistryIteratorEnterEntry. The iterator object keeps track of entries that have been recursed into previously to avoid loops.

Parameter mainPort: The main port obtained from IOMainPort(). Pass kIOMainPortDefault to look up the default main port.

Parameter plane: The name of an existing registry plane. Plane names are defined in IOKitKeys.h, eg. kIOServicePlane.

Parameter options: kIORegistryIterateRecursively may be set to recurse automatically into each entry as it is returned from IOIteratorNext calls on the registry iterator.

Parameter iterator: A created iterator handle, to be released by the caller when it has finished with it.

Returns: A kern_return_t error code.

Safety

plane Array TODO.

iterator must be a valid pointer.