objc2_io_kit

Struct IOHIDUserDevice

pub struct IOHIDUserDevice { /* private fields */ }

Apple’s documentation

Implementations

impl IOHIDUserDevice

pub unsafe fn with_properties( allocator: Option<&CFAllocator>, properties: &CFDictionary, options: IOOptionBits, ) -> Option<CFRetained<IOHIDUserDevice>>

Creates a virtual IOHIDDevice in the kernel.

The IOHIDUserDeviceRef represents a virtual IOHIDDevice. In order to create the device, the entitlement “com.apple.developer.hid.virtual.device” is required to validate the source of the device.

Parameter allocator: Allocator to be used during creation.

Parameter properties: Dictionary containing device properties indexed by keys defined in IOHIDKeys.h. At the bare minimum, the kIOHIDReportDescriptorKey key must be provided, where the value represents a CFData representation of the device’s report descriptor.

Parameter options: Options to be used when creating the device.

Returns: Returns a IOHIDUserDeviceRef on success.

Safety

properties generics must be of the correct type.

pub fn activate(&self)

Activates the IOHIDUserDevice object.

An IOHIDUserDevice object associated with a dispatch queue is created in an inactive state. The object must be activated in order to receive asynchronous events from the kernel.

A dispatch queue must be set via IOHIDUserDeviceSetDispatchQueue before activation.

An activated device must be cancelled via IOHIDUserDeviceCancel. All calls to “Register” functions should be done before activation and not after cancellation.

Calling IOHIDUserDeviceActivate on an active IOHIDUserDevice has no effect.

Parameter device: Reference to an IOHIDUserDevice.

pub fn cancel(&self)

Cancels the IOHIDUserDevice preventing any further invocation of its event handler block.

Cancelling prevents any further invocation of the event handler block for the specified dispatch queue, but does not interrupt an event handler block that is already in progress.

Explicit cancellation of the IOHIDUserDevice is required, no implicit cancellation takes place.

Calling IOHIDUserDeviceCancel on an already cancelled queue has no effect.

The IOHIDUserDeviceRef should only be released after the device has been cancelled, and the cancel handler has been called. This is to ensure all asynchronous objects are released. For example:

dispatch_block_t cancelHandler = dispatch_block_create(0, ^{ CFRelease(device); }); IOHIDUserDeviceSetCancelHandler(device, cancelHandler); IOHIDUserDeviceActivate(device); IOHIDUserDeviceCancel(device);

Parameter device: Reference to an IOHIDUserDevice

pub fn property(&self, key: &CFString) -> Option<CFRetained<CFType>>

Obtains a property from the device.

Parameter key: The property key.

Returns: Returns the property on success.

pub unsafe fn set_property(&self, key: &CFString, property: &CFType) -> bool

Sets a property on the device.

Parameter key: The property key.

Parameter value: The value of the property.

Returns: Returns true on success.

Safety

property should be of the correct type.

pub unsafe fn handle_report_with_time_stamp( &self, timestamp: u64, report: NonNull<u8>, report_length: CFIndex, ) -> IOReturn

Dispatches a report on behalf of the device.

Parameter device: Reference to a IOHIDUserDeviceRef.

Parameter timestamp: mach_absolute_time() based timestamp.

Parameter report: Buffer containing a HID report.

Parameter reportLength: The report buffer length.

Returns: Returns kIOReturnSuccess on success.

Safety

report must be a valid pointer.